# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Kalar Map
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PlaceCity = (typeof PlaceCity)[keyof typeof PlaceCity];

export const PlaceCity = {
  kalar: "kalar",
  kifre: "kifre",
  rizgari: "rizgari",
} as const;

export type PlaceCategory = (typeof PlaceCategory)[keyof typeof PlaceCategory];

export const PlaceCategory = {
  mosque: "mosque",
  school: "school",
  government: "government",
  hospital: "hospital",
  market: "market",
  university: "university",
  institute: "institute",
  shop: "shop",
  stadium: "stadium",
  park: "park",
  cemetery: "cemetery",
  hotel: "hotel",
  restaurant: "restaurant",
  cafe: "cafe",
  recreation: "recreation",
} as const;

export interface PlaceImage {
  id: number;
  placeId: number;
  url: string;
  caption?: string | null;
}

export interface Place {
  id: number;
  name: string;
  nameKurdish: string;
  city: PlaceCity;
  category: PlaceCategory;
  description?: string | null;
  phone?: string | null;
  latitude: number;
  longitude: number;
  images: PlaceImage[];
  createdAt: string;
}

export type CreatePlaceBodyCity =
  (typeof CreatePlaceBodyCity)[keyof typeof CreatePlaceBodyCity];

export const CreatePlaceBodyCity = {
  kalar: "kalar",
  kifre: "kifre",
  rizgari: "rizgari",
} as const;

export type CreatePlaceBodyCategory =
  (typeof CreatePlaceBodyCategory)[keyof typeof CreatePlaceBodyCategory];

export const CreatePlaceBodyCategory = {
  mosque: "mosque",
  school: "school",
  government: "government",
  hospital: "hospital",
  market: "market",
  university: "university",
  institute: "institute",
  shop: "shop",
  stadium: "stadium",
  park: "park",
  cemetery: "cemetery",
  hotel: "hotel",
  restaurant: "restaurant",
  cafe: "cafe",
  recreation: "recreation",
} as const;

export interface CreatePlaceBody {
  name: string;
  nameKurdish: string;
  city: CreatePlaceBodyCity;
  category: CreatePlaceBodyCategory;
  description?: string | null;
  phone?: string | null;
  latitude: number;
  longitude: number;
  imageUrl?: string | null;
}

export type UpdatePlaceBodyCity =
  (typeof UpdatePlaceBodyCity)[keyof typeof UpdatePlaceBodyCity];

export const UpdatePlaceBodyCity = {
  kalar: "kalar",
  kifre: "kifre",
  rizgari: "rizgari",
} as const;

export type UpdatePlaceBodyCategory =
  (typeof UpdatePlaceBodyCategory)[keyof typeof UpdatePlaceBodyCategory];

export const UpdatePlaceBodyCategory = {
  mosque: "mosque",
  school: "school",
  government: "government",
  hospital: "hospital",
  market: "market",
  university: "university",
  institute: "institute",
  shop: "shop",
  stadium: "stadium",
  park: "park",
  cemetery: "cemetery",
  hotel: "hotel",
  restaurant: "restaurant",
  cafe: "cafe",
  recreation: "recreation",
} as const;

export interface UpdatePlaceBody {
  name?: string;
  nameKurdish?: string;
  city?: UpdatePlaceBodyCity;
  category?: UpdatePlaceBodyCategory;
  description?: string | null;
  phone?: string | null;
  latitude?: number;
  longitude?: number;
}

export interface AddImageBody {
  url: string;
  caption?: string | null;
}

export type PlaceStatsByCity = { [key: string]: number };

export type PlaceStatsByCategory = { [key: string]: number };

export interface PlaceStats {
  totalPlaces: number;
  byCity: PlaceStatsByCity;
  byCategory: PlaceStatsByCategory;
}

export type ListPlacesParams = {
  city?: ListPlacesCity;
  category?: ListPlacesCategory;
  search?: string;
};

export type ListPlacesCity =
  (typeof ListPlacesCity)[keyof typeof ListPlacesCity];

export const ListPlacesCity = {
  kalar: "kalar",
  kifre: "kifre",
  rizgari: "rizgari",
} as const;

export type ListPlacesCategory =
  (typeof ListPlacesCategory)[keyof typeof ListPlacesCategory];

export const ListPlacesCategory = {
  mosque: "mosque",
  school: "school",
  government: "government",
  hospital: "hospital",
  market: "market",
  university: "university",
  institute: "institute",
  shop: "shop",
  stadium: "stadium",
  park: "park",
  cemetery: "cemetery",
  hotel: "hotel",
  restaurant: "restaurant",
  cafe: "cafe",
  recreation: "recreation",
} as const;