# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Kalar Map
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all places
 */
export const ListPlacesQueryParams = zod.object({
  city: zod.enum(["kalar", "kifre", "rizgari"]).optional(),
  category: zod
    .enum([
      "mosque",
      "school",
      "government",
      "hospital",
      "market",
      "university",
      "institute",
      "shop",
      "stadium",
      "park",
      "cemetery",
      "hotel",
      "restaurant",
      "cafe",
      "recreation",
    ])
    .optional(),
  search: zod.coerce.string().optional(),
});

export const ListPlacesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  nameKurdish: zod.string(),
  city: zod.enum(["kalar", "kifre", "rizgari"]),
  category: zod.enum([
    "mosque",
    "school",
    "government",
    "hospital",
    "market",
    "university",
    "institute",
    "shop",
    "stadium",
    "park",
    "cemetery",
    "hotel",
    "restaurant",
    "cafe",
    "recreation",
  ]),
  description: zod.string().nullish(),
  phone: zod.string().nullish(),
  latitude: zod.number(),
  longitude: zod.number(),
  images: zod.array(
    zod.object({
      id: zod.number(),
      placeId: zod.number(),
      url: zod.string(),
      caption: zod.string().nullish(),
    }),
  ),
  createdAt: zod.string(),
});
export const ListPlacesResponse = zod.array(ListPlacesResponseItem);

/**
 * @summary Create a new place
 */
export const CreatePlaceBody = zod.object({
  name: zod.string(),
  nameKurdish: zod.string(),
  city: zod.enum(["kalar", "kifre", "rizgari"]),
  category: zod.enum([
    "mosque",
    "school",
    "government",
    "hospital",
    "market",
    "university",
    "institute",
    "shop",
    "stadium",
    "park",
    "cemetery",
    "hotel",
    "restaurant",
    "cafe",
    "recreation",
  ]),
  description: zod.string().nullish(),
  phone: zod.string().nullish(),
  latitude: zod.number(),
  longitude: zod.number(),
  imageUrl: zod.string().nullish(),
});

/**
 * @summary Get a place by ID
 */
export const GetPlaceParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPlaceResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  nameKurdish: zod.string(),
  city: zod.enum(["kalar", "kifre", "rizgari"]),
  category: zod.enum([
    "mosque",
    "school",
    "government",
    "hospital",
    "market",
    "university",
    "institute",
    "shop",
    "stadium",
    "park",
    "cemetery",
    "hotel",
    "restaurant",
    "cafe",
    "recreation",
  ]),
  description: zod.string().nullish(),
  phone: zod.string().nullish(),
  latitude: zod.number(),
  longitude: zod.number(),
  images: zod.array(
    zod.object({
      id: zod.number(),
      placeId: zod.number(),
      url: zod.string(),
      caption: zod.string().nullish(),
    }),
  ),
  createdAt: zod.string(),
});

/**
 * @summary Update a place
 */
export const UpdatePlaceParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdatePlaceBody = zod.object({
  name: zod.string().optional(),
  nameKurdish: zod.string().optional(),
  city: zod.enum(["kalar", "kifre", "rizgari"]).optional(),
  category: zod
    .enum([
      "mosque",
      "school",
      "government",
      "hospital",
      "market",
      "university",
      "institute",
      "shop",
      "stadium",
      "park",
      "cemetery",
      "hotel",
      "restaurant",
      "cafe",
      "recreation",
    ])
    .optional(),
  description: zod.string().nullish(),
  phone: zod.string().nullish(),
  latitude: zod.number().optional(),
  longitude: zod.number().optional(),
});

export const UpdatePlaceResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  nameKurdish: zod.string(),
  city: zod.enum(["kalar", "kifre", "rizgari"]),
  category: zod.enum([
    "mosque",
    "school",
    "government",
    "hospital",
    "market",
    "university",
    "institute",
    "shop",
    "stadium",
    "park",
    "cemetery",
    "hotel",
    "restaurant",
    "cafe",
    "recreation",
  ]),
  description: zod.string().nullish(),
  phone: zod.string().nullish(),
  latitude: zod.number(),
  longitude: zod.number(),
  images: zod.array(
    zod.object({
      id: zod.number(),
      placeId: zod.number(),
      url: zod.string(),
      caption: zod.string().nullish(),
    }),
  ),
  createdAt: zod.string(),
});

/**
 * @summary Delete a place
 */
export const DeletePlaceParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Add an image to a place
 */
export const AddPlaceImageParams = zod.object({
  id: zod.coerce.number(),
});

export const AddPlaceImageBody = zod.object({
  url: zod.string(),
  caption: zod.string().nullish(),
});

/**
 * @summary Delete a place image
 */
export const DeletePlaceImageParams = zod.object({
  id: zod.coerce.number(),
  imageId: zod.coerce.number(),
});

/**
 * @summary Get place statistics by city and category
 */
export const GetPlaceStatsResponse = zod.object({
  totalPlaces: zod.number(),
  byCity: zod.record(zod.string(), zod.number()),
  byCategory: zod.record(zod.string(), zod.number()),
});