# lib/api-zod/src/generated/types/createPlaceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Kalar Map
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePlaceBodyCategory } from "./createPlaceBodyCategory";
import type { CreatePlaceBodyCity } from "./createPlaceBodyCity";

export interface CreatePlaceBody {
  name: string;
  nameKurdish: string;
  city: CreatePlaceBodyCity;
  category: CreatePlaceBodyCategory;
  description?: string | null;
  phone?: string | null;
  latitude: number;
  longitude: number;
  imageUrl?: string | null;
}