# lib/api-zod/src/generated/types/place.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Kalar Map
 * OpenAPI spec version: 0.1.0
 */
import type { PlaceCategory } from "./placeCategory";
import type { PlaceCity } from "./placeCity";
import type { PlaceImage } from "./placeImage";

export interface Place {
  id: number;
  name: string;
  nameKurdish: string;
  city: PlaceCity;
  category: PlaceCategory;
  description?: string | null;
  phone?: string | null;
  latitude: number;
  longitude: number;
  images: PlaceImage[];
  createdAt: string;
}