# lib/api-zod/src/generated/types/placeImage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Kalar Map
 * OpenAPI spec version: 0.1.0
 */

export interface PlaceImage {
  id: number;
  placeId: number;
  url: string;
  caption?: string | null;
}