# lib/api-zod/src/generated/types/placeStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Kalar Map
 * OpenAPI spec version: 0.1.0
 */
import type { PlaceStatsByCategory } from "./placeStatsByCategory";
import type { PlaceStatsByCity } from "./placeStatsByCity";

export interface PlaceStats {
  totalPlaces: number;
  byCity: PlaceStatsByCity;
  byCategory: PlaceStatsByCategory;
}