# lib/api-zod/src/generated/types/updatePlaceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Kalar Map
 * OpenAPI spec version: 0.1.0
 */
import type { UpdatePlaceBodyCategory } from "./updatePlaceBodyCategory";
import type { UpdatePlaceBodyCity } from "./updatePlaceBodyCity";

export interface UpdatePlaceBody {
  name?: string;
  nameKurdish?: string;
  city?: UpdatePlaceBodyCity;
  category?: UpdatePlaceBodyCategory;
  description?: string | null;
  phone?: string | null;
  latitude?: number;
  longitude?: number;
}